Add Makefile with expected-to-succeed "make check" rule.
The rule prepares test database and runs all tests found to succeed
on my personal local deploy. This is a fraction of all available
tests. The list of run tests should grow over time.

In addition to the "check" rule, which does both db preparation and
test run, there's a "check-prepared" rule that assumes a test
database was already prepared and a "prepare-test-db" which only
prepares the database.